requirements.txt
pandas
requests
google-generativeai
playwright
